update sam forward
Sam now returns only low-res logits, and expects the input image and prompts to be preprocessed

- Remove pre and post processing from sam arch
- Modify sam prediction to use sam forward
