+++ kornia/contrib/sam/__init__.py
@@ -1,3 +1,3 @@
-from kornia.contrib.sam.architecture import Sam, SamModelType
+from kornia.contrib.sam.model import Sam, SamModelType
 
 # from kornia.contrib.sam.predictor import SamPredictor
