+++ kornia/contrib/sam/architecture/__init__.py
@@ -1,5 +1,5 @@
 from kornia.contrib.sam.architecture.image_encoder import ImageEncoderViT
 from kornia.contrib.sam.architecture.mask_decoder import MaskDecoder
 from kornia.contrib.sam.architecture.prompt_encoder import PromptEncoder
-from kornia.contrib.sam.architecture.sam import Sam, SamModelType
+from kornia.contrib.sam.architecture.sam import Sam
 from kornia.contrib.sam.architecture.transformer import TwoWayTransformer
